Extracted ExecutionContext module from Executor
